Feat: Add APIData.ts - API data layout 파일에서 빼서 새로운 파일에 넣어줬습니다.
